restructuring: put everything in calc-lib
