+++ crates/calc-lib/src/functions/minus.rs
