+++ crates/calc-lib/src/functions/mult.rs
