+++ crates/calc-lib/src/utils/integer_utils.rs
