+++ crates/calc-lib/src/utils/matrix_utils.rs
