+++ crates/calc-lib/src/utils/plot_utils.rs
